[ci skip] Create .def from .def.dist
We export aliases for external libpcre.

This doesn't work though, since fileinfo, for instance, refers to
`pcre2_match_8`, which is not exported under that name.  If we remove
`HAVE_BUNDLED_PCRE` from php_compat.h, same problem.  If we use #if 1
there, then there are link issues with the static exts, which don't get
the export aliases.  We would need some way to detect whether the
currently compiled object is supposed to be statically linked or not.