Increase Opcache log verbosity level on AppVeyor
The default opcache.log_verbosity_level lets some tests pass, even
though they did not test what was supposed to be tested.  For instance,
the failing bug76275.phpt (reported as bug #77992) did not fail on
AppVeyor, since it simply ignored the file cache setting, and thus
did not serve as meaningful regression test.

We therefore increase the verbosity level to also include warnings in
the output.